refactor(lex_gen): extract shared FreezedModel base for object/record/input/output
