+++ packages/lex_gen/lib/src/services/object/lex_input.dart
@@ -4,20 +4,10 @@
 
 // Project imports:
 import '../gen_context.dart';
-import 'lex_property.dart';
 import 'lex_type.dart';
 import 'utils.dart';
 
-final class LexInput extends GeneratableType {
-  @override
-  final String lexiconId;
-  @override
-  final String defName;
-
-  final String name;
-  final String? description;
-  final List<LexProperty> properties;
-
+final class LexInput extends FreezedModel {
   final bool bytes;
   final String? encoding;
 
@@ -27,11 +17,11 @@ final class LexInput extends GeneratableType {
   LexTypeState get state => LexTypeState.input;
 
   const LexInput({
-    required this.lexiconId,
-    required this.defName,
-    required this.name,
-    this.description,
-    required this.properties,
+    required super.lexiconId,
+    required super.defName,
+    required super.name,
+    super.description,
+    required super.properties,
     this.bytes = false,
     this.ref,
     this.encoding,
@@ -50,11 +40,6 @@ final class LexInput extends GeneratableType {
   @override
   bool isBytes() {
     return bytes;
-  }
-
-  @override
-  List<LexProperty> getProperties() {
-    return properties;
   }
 
   @override
