+++ packages/lex_gen/lib/src/services/object/lex_record.dart
@@ -5,35 +5,20 @@
 // Project imports:
 import '../gen_context.dart';
 import '../rule.dart' as rule;
-import 'lex_property.dart';
 import 'lex_type.dart';
 import 'utils.dart';
 
-final class LexRecord extends GeneratableType {
-  @override
-  final String lexiconId;
-  @override
-  final String defName;
-
-  final String name;
-  final String? description;
-  final List<LexProperty> properties;
-
-  @override
-  List<LexProperty> getProperties() {
-    return properties;
-  }
+final class LexRecord extends FreezedModel {
+  const LexRecord({
+    required super.lexiconId,
+    required super.defName,
+    required super.name,
+    super.description,
+    required super.properties,
+  });
 
   @override
   LexTypeState get state => LexTypeState.record;
-
-  const LexRecord({
-    required this.lexiconId,
-    required this.defName,
-    required this.name,
-    this.description,
-    required this.properties,
-  });
 
   @override
   String getTypeName() {
@@ -51,17 +36,11 @@ final class LexRecord extends GeneratableType {
       description: description,
       properties: properties,
       typeDefaultId: id,
-      validateMethod: _getValidateMethod(id),
+      validateMethod: buildValidateMethod(
+        id,
+        includeSubscription: false,
+        includeMainAlias: false,
+      ),
     );
-  }
-
-  String _getValidateMethod(final String id) {
-    final buffer = StringBuffer();
-    buffer.writeln('static bool validate(final Map<String, dynamic> object) {');
-    buffer.writeln("  if (!object.containsKey('\\\$type')) return false;");
-    buffer.writeln("  return object['\\\$type'] == '$id';");
-    buffer.writeln('}');
-
-    return buffer.toString();
   }
 }
